Backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pymongo>=4.9.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
pydantic[email]>=2.9.0
Pillow>=10.4.0
requests>=2.32.0
python-multipart>=0.0.12
slowapi>=0.1.9
